requirements.txt
PyYAML>=6.0
pymysql>=1.1.0
sqlalchemy>=2.0.0
pytest>=7.4.0
colorama>=0.4.6
tabulate>=0.9.0
